# Extract worklet logic from value unpacker