<!-- Thanks for submitting a pull request! We appreciate you spending the time to work on these changes. Please follow the template so that the reviewers can easily understand what the code changes affect. -->

## Summary

This PR moves the logic behind creation of worklets from JS `valueUnpacker` directly into C++ `ShareableWorklet::toJSValue`.

It's a step towards removing `valueUnpacker` which has been causing some problems recently (see #5660).

## Test plan

<!-- Provide a minimal but complete code snippet that can be used to test out this change along with instructions how to run it and a description of the expected behavior. -->