@tomekzaw 2024-04-02
Some performance measurements:

main debug 136360 us
main release 86487 us

this PR debug 397071 us
this PR release 242305 us

😢