@tjzel 2024-04-02
Haven't had the time to check it yet, but did you verify that the caches are actually utilized? Maybe that's the reason for performance issues.