@tomekzaw 2024-04-03
@tjzel You are right, it looks like after the changes the cache does not work as expected.

Before:
```
worklet hash: creation time
17031627990684: 109 us
17031627990684: 5 us
17031627990684: 3 us
17031627990684: 3 us
17031627990684: 3 us
17031627990684: 3 us
17031627990684: 3 us
17031627990684: 3 us
```

After:
```
worklet hash: creation time
17031627990684: 123 us
17031627990684: 110 us
17031627990684: 110 us
17031627990684: 108 us
17031627990684: 107 us
17031627990684: 107 us
17031627990684: 113 us
17031627990684: 106 us
```