@tomekzaw 2024-04-04
After:
this PR debug: 122862 us
this PR release: 82824 us

So it's faster 🥳