@tomekzaw 2024-04-04
## CWCT (cumulative worklet creation time)

```cpp
#include <iostream>
#include <sstream>
#include <chrono>

using namespace std::chrono;

auto start = high_resolution_clock::now();
// ShareableWorklet::toJSValue
auto end = high_resolution_clock::now();
auto duration = duration_cast<microseconds>(end - start);
static unsigned long acc = 0;
acc += duration.count();
std::cout << static_cast<unsigned long long>(workletHash.asNumber()) << ": " << acc << " us" << std::endl;
```

| | main | this PR |
|-|:-:|:-:|
| **debug** | 122455 us | 122862 us |
| **release** | 82094 us | 82824 us |

So it's just a 1% slower 🥳