@tomekzaw 2024-04-04
## WIT (worklet invocation time)

```ts
function worklet() {
  'worklet';
  // do nothing
}
runOnUI(() => {
  const start = performance.now();
  for (let i = 0; i < 10000000; i++) {
    worklet();
  }
  const end = performance.now();
  console.log(end - start, 'ms');
})();
```

| | main | this PR |
|-|:-:|:-:|
| **debug** | 373 us | 371 us |
| **release** | 347 us | 347 us |